% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{arrange_path}
\alias{arrange_path}
\title{Helper function to reorder the values in the input flight dataframe.}
\usage{
arrange_path(df)
}
\arguments{
\item{df}{A two-column dataframe - The start and end airports of flights.}
}
\value{
A two-column dataframe
}
\description{
Helper function to reorder the values in the input flight dataframe.
}
\examples{
arrange_path(sample_trips)

}
