% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_trips}
\alias{sample_trips}
\title{Sample Trip Dataset}
\format{
A two-column dataframe with start and end airports of flights.
}
\usage{
sample_trips
}
\description{
A two-column dataframe with start and end airports of flights. The dataframe can be readily used as input for the main plot function.
}
\keyword{datasets}
