% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world}
\alias{world}
\title{World Map Vector Data}
\format{
A simple feature dataframe.
}
\usage{
world
}
\description{
A simple feature dataframe generated from world land and earth shapefile.
}
\keyword{datasets}
