% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airports}
\alias{airports}
\title{The Information of All Aiports}
\format{
A dataframe with basic information of airports all over the world
}
\usage{
airports
}
\description{
A dataframe with basic information of airports all over the world
}
\keyword{datasets}
