% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_trips}
\alias{sample_trips}
\title{Sample Trip Dataset}
\format{
Dataframe
\describe{
\item{Departure}{IATA code for departure airports}
\item{Arrival}{IATA code for arrival airports.}
}
}
\usage{
sample_trips
}
\description{
A two-column dataframe with start and end airports of flights. The dataframe can be readily used as input for the main plot function.
}
\examples{
  sample_trips
}
\keyword{datasets}
